+++ astropy/coordinates/attributes.py
@@ -186,6 +186,10 @@ class TimeAttribute(Attribute):
                                                                value, err))
             converted = True
 
+        # Set attribute as read-only for arrays (not allowed by numpy
+        # for array scalars)
+        if out.shape:
+            out.writeable = False
         return out, converted